Removed the byte_mod.f90, grib_info_mod.f90, cosmo_data_mod.f90 footprint from the model_mod and removed  their reference in the path_names* files.
There seems to be a bug in the determination of the vertical layer index.
When model_mod_check tries to interpolate at exactly layer 50 (of 50) it
generates another index of 51 ... which causes a core dump.
The algorithm to determine the layers and their fractions needs more work.

The dart_to_cosmo program is underway but is not done.