parflow3_9: handling of CMake input -DPARFLOW_ENABLE_SLURM
`PARFLOW_ENABLE_SLURM` needs to be `FALSE` for now.

Enabling SLURM results in an "undefined" Error for
`slurm_get_rem_time` in `solver_richards.c`. Probable cause: the SLURM
library is not linked correctly.

As SLURM is turned off for `parflow3_2`, this behavior is hereby
copied for `parflow3_9` until further development.