intf_DA: set `cosmo_input_suffix` from input to `cosmo_init`
Before, `cosmo_input_suffix` was set in `init_parallel_pdaf`, where it
was slightly misplaced. The setting of this suffix for the COSMO input
files `INPUT_IO*` is best part of `cosmo_init()` in `wrapper_tsmp`.

In `wrapper_tsmp` we can - as another pro - make use of the definition
of `model` in `read_enkfpar`.