TSMP-PDAF update (#223)
* intf_DA: update MPI-use-command and add `abort_parallel` routine

* intf_DA: remove `dim_obs_f` from `mod_assimilation`

`dim_obs_f` is the dimension of the full observation vector. This
information is already saved in `dim_obs.

Setting `dim_obs_f` is removed. Instead, a check of the sum of
PE-local observation vectors vs `dim_obs` is introduced.

* intf_DA: `local_range` renamed to `cradius`

- https://pdaf.awi.de/trac/wiki/ReleaseNotes, Version 2.1
- `local_range` kept as command line input for backward compatibility

* intf_DA: `srange` renamed to `sradius`

- https://pdaf.awi.de/trac/wiki/ReleaseNotes, Version 2.1
- `srange` kept as command line input for backward compatibility

* intf_DA: remove unused variable `covartype`

* intf_DA: remove unused SEEK variables `int_rediag`, `epsilon`

* intf_DA: remove SEEK from comments

* intf_DA: include variables for (L)(K)NETF, PF, 3D-Var

* intf_DA: update `init_pdaf_info` for (L)(K)NETF, PF, 3D-Var

* intf_DA: defaults in `init_pdaf` for (L)(K)NETF, PF, 3D-Var

* intf_DA: move `abort_parallel` to `mod_parallel_pdaf`

* intf_DA: add `IMPLICIT NONE` in `mod_parallel_model`

* intf_DA: move `mod_parallel_model` into `mod_parallel_pdaf`

* intf_DA: update software name in copyright/documentation

- `TerrSysMP` -> `TSMP`

* intf_DA: refactor `pdaf_terrsysmp.F90` use-statements

* intf_DA: move five variables from `mod_parallel_pdaf`->`mod_tsmp`

Variables:
- `t_start`, `model`, `tcycle`, `tstartcycle`, `total_steps`

Reason: These variables are not related to parallelization

* intf_DA: always check `nreal` against `n_modeltasks`

* doc: parallelization input updates

* intf_DA: MPI consistency check for nproc inputs

* intf_DA: `CLMFIVE` changed to `CLMSA`

* intf_DA: set `cosmo_input_suffix` from input to `cosmo_init`

Before, `cosmo_input_suffix` was set in `init_parallel_pdaf`, where it
was slightly misplaced. The setting of this suffix for the COSMO input
files `INPUT_IO*` is best part of `cosmo_init()` in `wrapper_tsmp`.

In `wrapper_tsmp` we can - as another pro - make use of the definition
of `model` in `read_enkfpar`.

* intf_DA: rename `statcomm` to `COMM_couple_clm`

* intf_DA: rename `da_comm_clm` to `COMM_model_clm`

* intf_DA: introduce `COMM_model_pfl`

`COMM_model_pfl` is set to `COMM_model` in `init_parallel_pdaf`.

For `PARFLOW_STAND_ALONE`, `COMM_model_pfl` is turned into a
C-communicator-handle in `enkfparflowinit` and handed over to AMPS.

* intf_DA: rename `da_comm` to `COMM_model_oas`

* intf_DA: remove environment variable `use_comm_da`

* doc updates