intf_DA: PE configuration output with grep-hook `Pconf`
This way the lines of this table can easily be "grepped".

Source:
https://github.com/larsnerger/NEMO-PDAF/blob/main/MY_SRC_pdaf/parallel_pdaf.F90

Problem: Standard output does not follow the MPI-Barriers.

The OS first writes into process-specific buffers, which are later
added to the standard output and the OS does not know the order of the
MPI-Barriers and therefore sometimes does not follow it.